requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-chroma
chromadb
tiktoken
requests